--- a/test/JDBC/expected/BABEL-SP_FKEYS-dep-vu-verify.out
+++ b/test/JDBC/expected/BABEL-SP_FKEYS-dep-vu-verify.out
@@ -1,3 +1,4 @@
+
 exec babel_sp_fkeys_dep_vu_prepare_p1
 go
 ~~START~~