--- a/test/JDBC/expected/BABEL-SP_FKEYS-vu-verify.out
+++ b/test/JDBC/expected/BABEL-SP_FKEYS-vu-verify.out
@@ -1,3 +1,4 @@
+
 use babel_sp_fkeys_vu_prepare_db1
 go