--- a/test/JDBC/input/BABEL-SP_FKEYS-dep-vu-verify.sql
+++ b/test/JDBC/input/BABEL-SP_FKEYS-dep-vu-verify.sql
@@ -1,3 +1,3 @@
--- sla 7000000
+
 exec babel_sp_fkeys_dep_vu_prepare_p1
 go