--- a/test/JDBC/upgrade/13_6/schedule
+++ b/test/JDBC/upgrade/13_6/schedule
@@ -299,3 +299,5 @@ sys-sql_modules_before-14_7-or-15_2
 sys-all_sql_modules_before-14_5
 BABEL_OBJECT_ID
 test_windows_login_before_15_2
+BABEL-SP_FKEYS
+BABEL-SP_FKEYS-dep