--- a/test/JDBC/upgrade/14_5/schedule
+++ b/test/JDBC/upgrade/14_5/schedule
@@ -323,3 +323,5 @@ sys-all_sql_modules_before-14_7-or-15_2
 sys-sql_modules_before-14_7-or-15_2
 BABEL_OBJECT_ID
 test_windows_login_before_15_2
+BABEL-SP_FKEYS
+BABEL-SP_FKEYS-dep