… test/JDBC/upgrade/14_6/schedule
… test/JDBC/upgrade/14_7/schedule
… test/JDBC/upgrade/15_1/schedule
… test/JDBC/upgrade/latest/schedule